separate utils in input controllers by domain
