+++ railgun/controllers/inputs/ingress_utils.go
@@ -2,48 +2,14 @@ package inputs
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/go-logr/logr"
 	"gopkg.in/yaml.v2"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-const configSecretName = "kong-config"
-
-// getOrCreateConfigSecret finds or creates the secret which houses the combined ingress of the cluster for
-// eventual parsing and emitting to the Kong Admin API on the proxy instances.
-func getOrCreateConfigSecret(ctx context.Context, c client.Client, ns string) (*corev1.Secret, bool, error) {
-	secret := new(corev1.Secret)
-	if err := c.Get(ctx, types.NamespacedName{Namespace: ns, Name: configSecretName}, secret); err != nil {
-		if errors.IsNotFound(err) {
-			secret.SetName(configSecretName)
-			secret.SetNamespace(ns)
-			if err := c.Create(ctx, secret); err != nil {
-				return nil, false, err
-			}
-			return secret, true, nil
-		} else {
-			return nil, false, err
-		}
-	}
-	if secret.Data == nil {
-		secret.Data = make(map[string][]byte)
-	}
-	return secret, false, nil
-}
-
-// keyFor provides the string key that should be used to store the YAML contents of an object in the
-// Kong configuration secret.
-func keyFor(obj runtime.Object, nsn types.NamespacedName) string {
-	gvk := obj.GetObjectKind().GroupVersionKind()
-	return fmt.Sprintf("%s-%s-%s-%s-%s", gvk.Group, gvk.Version, gvk.Kind, nsn.Namespace, nsn.Name)
-}
 
 // isManaged verifies whether an Ingress resource is managed by Kong controllers by verifying the
 // annotations of the object.
@@ -88,7 +54,7 @@ func storeIngressUpdates(ctx context.Context, c client.Client, log logr.Logger, 
 		log.Info("kong configuration did not exist, was created successfully", "namespace", nsn.Namespace, "ingress", nsn.Name)
 		return ctrl.Result{Requeue: true}, nil
 	}
-	log.Info("kong configuration secret found", "namespace", nsn.Namespace, "name", configSecretName)
+	log.Info("kong configuration secret found", "namespace", nsn.Namespace, "name", ConfigSecretName)
 
 	// get the storage key for this ingress object
 	// TODO: patch instead of update for perf
@@ -102,7 +68,7 @@ func storeIngressUpdates(ctx context.Context, c client.Client, log logr.Logger, 
 		return ctrl.Result{}, err
 	}
 
-	log.Info("kong configuration patched (TODO: not doing a PATCH yet doing an UPDATE!)", "namespace", nsn.Namespace, "name", configSecretName)
+	log.Info("kong configuration patched (TODO: not doing a PATCH yet doing an UPDATE!)", "namespace", nsn.Namespace, "name", ConfigSecretName)
 	return ctrl.Result{}, nil
 }
